+++ pkg/profile/builders.go
@@ -1,6 +1,10 @@
 package profile
 
-import profilesv1 "github.com/weaveworks/profiles/api/v1alpha1"
+import (
+	profilesv1 "github.com/weaveworks/profiles/api/v1alpha1"
+
+	"github.com/weaveworks/pctl/pkg/profile/artifact"
+)
 
 const (
 	// KUSTOMIZE is a kustomization builder type
@@ -9,16 +13,9 @@ const (
 	CHART
 )
 
-// BuilderConfig defines some common configuration values for builders.
-type BuilderConfig struct {
-	GitRepositoryName      string
-	GitRepositoryNamespace string
-	RootDir                string
-}
-
 // Builder can build an artifacts from an installation and a profile artifact.
 //go:generate counterfeiter -o fakes/builder_maker.go . Builder
 type Builder interface {
 	// Build a single artifact from a profile artifact and installation.
-	Build(artifact profilesv1.Artifact, installation profilesv1.ProfileInstallation, definition profilesv1.ProfileDefinition) ([]Artifact, error)
+	Build(artifact profilesv1.Artifact, installation profilesv1.ProfileInstallation, definition profilesv1.ProfileDefinition) ([]artifact.Artifact, error)
 }
