+++ pkg/profile/chart/builder.go
@@ -1,4 +1,4 @@
-package profile
+package chart
 
 import (
 	"fmt"
@@ -10,23 +10,32 @@ import (
 	sourcev1 "github.com/fluxcd/source-controller/api/v1beta1"
 	profilesv1 "github.com/weaveworks/profiles/api/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/weaveworks/pctl/pkg/profile/artifact"
 )
 
-// ChartBuilder will build helm chart resources.
-type ChartBuilder struct {
-	BuilderConfig
+// Config defines some common configuration values for builders.
+type Config struct {
+	GitRepositoryName      string
+	GitRepositoryNamespace string
+	RootDir                string
+}
+
+// Builder will build helm chart resources.
+type Builder struct {
+	Config
 }
 
 // Build a single artifact from a profile artifact and installation.
-func (c *ChartBuilder) Build(artifact profilesv1.Artifact, installation profilesv1.ProfileInstallation, definition profilesv1.ProfileDefinition) ([]Artifact, error) {
-	a := Artifact{Name: artifact.Name}
-	helmRelease := c.makeHelmRelease(artifact, installation, definition.Name)
+func (c *Builder) Build(att profilesv1.Artifact, installation profilesv1.ProfileInstallation, definition profilesv1.ProfileDefinition) ([]artifact.Artifact, error) {
+	a := artifact.Artifact{Name: att.Name}
+	helmRelease := c.makeHelmRelease(att, installation, definition.Name)
 	a.Objects = append(a.Objects, helmRelease)
-	if artifact.Chart.Path != "" {
+	if att.Chart.Path != "" {
 		if c.GitRepositoryNamespace == "" && c.GitRepositoryName == "" {
 			return nil, fmt.Errorf("in case of local resources, the flux gitrepository object's details must be provided")
 		}
-		helmRelease.Spec.Chart.Spec.Chart = filepath.Join(c.RootDir, "artifacts", artifact.Name, artifact.Chart.Path)
+		helmRelease.Spec.Chart.Spec.Chart = filepath.Join(c.RootDir, "artifacts", att.Name, att.Chart.Path)
 		branch := installation.Spec.Source.Branch
 		if installation.Spec.Source.Tag != "" {
 			branch = installation.Spec.Source.Tag
@@ -34,16 +43,16 @@ func (c *ChartBuilder) Build(artifact profilesv1.Artifact, installation profiles
 		a.RepoURL = installation.Spec.Source.URL
 		a.SparseFolder = definition.Name
 		a.Branch = branch
-		a.PathsToCopy = append(a.PathsToCopy, artifact.Chart.Path)
+		a.PathsToCopy = append(a.PathsToCopy, att.Chart.Path)
 	}
-	if artifact.Chart.URL != "" {
-		helmRepository := c.makeHelmRepository(artifact.Chart.URL, artifact.Chart.Name, installation)
+	if att.Chart.URL != "" {
+		helmRepository := c.makeHelmRepository(att.Chart.URL, att.Chart.Name, installation)
 		a.Objects = append(a.Objects, helmRepository)
 	}
-	return []Artifact{a}, nil
+	return []artifact.Artifact{a}, nil
 }
 
-func (c *ChartBuilder) makeHelmRelease(artifact profilesv1.Artifact, installation profilesv1.ProfileInstallation, definitionName string) *helmv2.HelmRelease {
+func (c *Builder) makeHelmRelease(artifact profilesv1.Artifact, installation profilesv1.ProfileInstallation, definitionName string) *helmv2.HelmRelease {
 	var helmChartSpec helmv2.HelmChartTemplateSpec
 	if artifact.Chart.Path != "" {
 		helmChartSpec = c.makeGitChartSpec(path.Join(installation.Spec.Source.Path, artifact.Chart.Path))
@@ -70,7 +79,7 @@ func (c *ChartBuilder) makeHelmRelease(artifact profilesv1.Artifact, installatio
 	return helmRelease
 }
 
-func (c *ChartBuilder) makeHelmRepository(url string, name string, installation profilesv1.ProfileInstallation) *sourcev1.HelmRepository {
+func (c *Builder) makeHelmRepository(url string, name string, installation profilesv1.ProfileInstallation) *sourcev1.HelmRepository {
 	return &sourcev1.HelmRepository{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      c.makeHelmRepoName(name, installation),
@@ -86,13 +95,25 @@ func (c *ChartBuilder) makeHelmRepository(url string, name string, installation 
 	}
 }
 
-func (c *ChartBuilder) makeHelmRepoName(name string, installation profilesv1.ProfileInstallation) string {
+func (c *Builder) makeHelmRepoName(name string, installation profilesv1.ProfileInstallation) string {
 	repoParts := strings.Split(installation.Spec.Source.URL, "/")
 	repoName := repoParts[len(repoParts)-1]
 	return join(installation.Name, repoName, name)
 }
 
-func (c *ChartBuilder) makeGitChartSpec(path string) helmv2.HelmChartTemplateSpec {
+func makeArtifactName(name string, installationName, definitionName string) string {
+	// if this is a nested artifact, it's name contains a /
+	if strings.Contains(name, "/") {
+		name = filepath.Base(name)
+	}
+	return join(installationName, definitionName, name)
+}
+
+func join(s ...string) string {
+	return strings.Join(s, "-")
+}
+
+func (c *Builder) makeGitChartSpec(path string) helmv2.HelmChartTemplateSpec {
 	return helmv2.HelmChartTemplateSpec{
 		Chart: path,
 		SourceRef: helmv2.CrossNamespaceObjectReference{
@@ -103,7 +124,7 @@ func (c *ChartBuilder) makeGitChartSpec(path string) helmv2.HelmChartTemplateSpe
 	}
 }
 
-func (c *ChartBuilder) makeHelmChartSpec(chart string, version string, installation profilesv1.ProfileInstallation) helmv2.HelmChartTemplateSpec {
+func (c *Builder) makeHelmChartSpec(chart string, version string, installation profilesv1.ProfileInstallation) helmv2.HelmChartTemplateSpec {
 	return helmv2.HelmChartTemplateSpec{
 		Chart: chart,
 		SourceRef: helmv2.CrossNamespaceObjectReference{
