+++ pkg/profile/chart/builder_test.go
@@ -1,4 +1,4 @@
-package profile_test
+package chart_test
 
 import (
 	helmv2 "github.com/fluxcd/helm-controller/api/v2beta1"
@@ -9,20 +9,32 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
-	"github.com/weaveworks/pctl/pkg/profile"
+	"github.com/weaveworks/pctl/pkg/profile/artifact"
+	"github.com/weaveworks/pctl/pkg/profile/chart"
 )
 
-var _ = Describe("ChartBuilder", func() {
+var _ = Describe("Builder", func() {
 	var (
 		profileName            string
 		profileURL             string
 		profilePath            string
-		artifact               profilesv1.Artifact
+		partifact              profilesv1.Artifact
 		pSub                   profilesv1.ProfileInstallation
 		pDef                   profilesv1.ProfileDefinition
 		rootDir                string
 		gitRepositoryName      string
 		gitRepositoryNamespace string
+		profileName1           = "weaveworks-nginx"
+		namespace              = "default"
+		profileSubAPIVersion   = "weave.works/v1alpha1"
+		profileSubKind         = "ProfileInstallation"
+	)
+
+	var (
+		profileTypeMeta = metav1.TypeMeta{
+			Kind:       profileSubKind,
+			APIVersion: profileSubAPIVersion,
+		}
 	)
 	BeforeEach(func() {
 		profileName = "test-profile"
@@ -42,7 +54,7 @@ var _ = Describe("ChartBuilder", func() {
 				},
 			},
 		}
-		artifact = profilesv1.Artifact{
+		partifact = profilesv1.Artifact{
 			Name: "dokuwiki",
 			Chart: &profilesv1.Chart{
 				URL:     "https://charts.bitnami.com/bitnami",
@@ -62,7 +74,7 @@ var _ = Describe("ChartBuilder", func() {
 				ProfileDescription: profilesv1.ProfileDescription{
 					Description: "foo",
 				},
-				Artifacts: []profilesv1.Artifact{artifact},
+				Artifacts: []profilesv1.Artifact{partifact},
 			},
 		}
 		rootDir = "root-dir"
@@ -72,14 +84,14 @@ var _ = Describe("ChartBuilder", func() {
 
 	Context("Build", func() {
 		It("creates an artifact from an install and a profile definition", func() {
-			chartBuilder := &profile.ChartBuilder{
-				BuilderConfig: profile.BuilderConfig{
+			chartBuilder := &chart.Builder{
+				Config: chart.Config{
 					GitRepositoryName:      gitRepositoryName,
 					GitRepositoryNamespace: gitRepositoryNamespace,
 					RootDir:                rootDir,
 				},
 			}
-			artifacts, err := chartBuilder.Build(artifact, pSub, pDef)
+			artifacts, err := chartBuilder.Build(partifact, pSub, pDef)
 			Expect(err).NotTo(HaveOccurred())
 			helmRelease := &helmv2.HelmRelease{
 				TypeMeta: metav1.TypeMeta{
@@ -117,7 +129,7 @@ var _ = Describe("ChartBuilder", func() {
 					URL: "https://charts.bitnami.com/bitnami",
 				},
 			}
-			expected := profile.Artifact{
+			expected := artifact.Artifact{
 				Objects: []runtime.Object{helmRelease, helmRepository},
 				Name:    "dokuwiki",
 			}
@@ -125,12 +137,12 @@ var _ = Describe("ChartBuilder", func() {
 		})
 		When("git-repository-name and git-repository-namespace aren't defined", func() {
 			It("returns an error", func() {
-				chartBuilder := &profile.ChartBuilder{
-					BuilderConfig: profile.BuilderConfig{
+				chartBuilder := &chart.Builder{
+					Config: chart.Config{
 						RootDir: rootDir,
 					},
 				}
-				artifact = profilesv1.Artifact{
+				partifact = profilesv1.Artifact{
 					Name: "local-artifact",
 					Chart: &profilesv1.Chart{
 						Path: "nginx/chart",
@@ -148,10 +160,10 @@ var _ = Describe("ChartBuilder", func() {
 						ProfileDescription: profilesv1.ProfileDescription{
 							Description: "foo",
 						},
-						Artifacts: []profilesv1.Artifact{artifact},
+						Artifacts: []profilesv1.Artifact{partifact},
 					},
 				}
-				_, err := chartBuilder.Build(artifact, pSub, pDef)
+				_, err := chartBuilder.Build(partifact, pSub, pDef)
 				Expect(err).To(MatchError("in case of local resources, the flux gitrepository object's details must be provided"))
 			})
 		})
