+++ pkg/profile/kustomize/builder_suite_test.go
@@ -1,0 +1,13 @@
+package kustomize_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestBuilder(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Builder Suite")
+}
